requirements.txt
twitchio==2.10.0
python-dotenv
aiohttp
discord.py
